Move afterEach tests into integration test dir
The afterEach.test.js file would spawn a test runner for each test
case. I needed to have it spawn jest as well, but also run the
remaining tests in both mocha and jest. If this file was not moved
out of the test folder, both mocha and jest would run the same test
file that spawned both mocha and jest.

This way, we only use mocha to run the integration test suite and
have it spawn both itself and jest for each test case when we get
to implement jest support in that file later.